feat: Reorganize and describe math funcs
- Move all math-related functions to a separate file
- Deprecate `logN()` (single-argument form only), `round()`,
  `squareRoot()`, `truncate()`
- Add JSDoc comments for all math-related functions
