+++ src/index.d.ts
@@ -1,4 +1,5 @@
 import './global';
 export * from './kolmafia';
 export * from './kolmafia-cargo-cultist-shorts';
+export * from './kolmafia-math';
 export * from './kolmafia-modifier';
